Add files via upload
Elevate Car Detail Webpage

This project is a basic HTML webpage created for a community college assignment. The webpage uses foundational HTML tags such as <html>, <head>, <title>, and <body>. It features:

A main heading and a secondary heading that introduces the business.
A paragraph that describes the car detailing services offered by Elevate Car Detail.
A hyperlink to the official website (https://www.elevatecardetail.com).
An unordered list that outlines some of the key services.